Remove events filtering
First, I already decide what events to send in the Webhook setup done in GitLab.
Having also this filtering might lead to problem with people that want to set up
a Webhook using a common link but not having access to the Mattermost receiver.
Also since the default configuration was True for all, it was definitely only a
choice done in GitLab.
Second, having the setup eventually made also using ENV will cause more troubles
and my view is using a JSON to configure everything, easy and plain.